Generate the final custom table
Rather than reproduce all the content, in this case we take the existing content. It's possible that we could have done this for the other HTML tables, but they'll never need to change (whereas this one might).

(This fixes an issue which was accidentally closed in the old repository. I haven't bothered reopening and transferring it.)