Provide guidance about creating issues
In particular, request that contributors hold off mentioning
features that aren't in the version we're currently standardising.

We don't want to discourage contributors in general, and if we miss
something we definitely want to know - but having huge numbers of
issues for features we aren't ready to standardize ends up making it
harder to find issues we *do* want to address.

Example issue: #296 - we'll close that once this change is in.