Validate that lists are preceded by a blank line
A list item can also be preceded by another list item with the same bullet type, or a previous nested list item.

Fixes #436

While I would *prefer* to do this elsewhere, having it here is better than not having it at all. We can try to find a good Markdown validator (that copes with everything we do) later.
This is really crude, but did help discover a lot of broken places in the spec.