configure markdown lint (#522)
* configure markdown lint

First configuration of markdown lint. We can determine which warnings we want to enable or disable from a first test run.

* disable a few items

- Disable long line lengths. (We use hard breaks for paragraphs)
- Disable inline HTML.

* add default problem matcher

* Disable rules currently generating warngs.

A couple note on this commit:

MD034 is for "raw URL link". The standard contains the link to the ECMA site, which is intentional. If it helps the word converter, that can be formatted using markdown's `[]()` syntax.

MD30 and MD31 will be the two biggest jobs.

Also, some of the warnings are generated from the markdown source in the test project. I think the right move will be to update those test sources once the associated rules are enabled.

* missed a few...