Format endnote and endexample markers (#534)
* format *end note*

The *end note* marker for a note should be at the end of a single paragraph note. If a note spans multiple paragraphs (including notes that contain examples or lists), the *end note* marker should be its own paragraph.

* Format *end example*

The *end example* marker for a note should be at the end of a single paragraph note. If a note spans multiple paragraphs (including notes that contain examples or lists), the *end example* marker should be its own paragraph.

* fix lint warnings.

* fix markdown converter issues

One example and one note had their final paragraph not included in the block quote.

* add link to comment on code change

This code change may be wrong. See comment.